refactor: rename of card date input component to index
